removed comment and moved signToken to utils
