+++ server/utils/index.js
@@ -1,4 +1,5 @@
-const {signToken, validateSchema} = require("./validation")
+const validateSchema = require("./validation");
+const signToken = require("./signToken")
 
 module.exports = {
   validateSchema,
