+++ server/utils/validation/index.js
@@ -1,7 +1,15 @@
-const validateSchema = require("./schemaValidation");
-const signToken = require("./signToken");
+const Joi = require("joi");
 
-module.exports = {
-  validateSchema,
-  signToken,
-};
+const validateSchema = Joi.object().keys({
+  email: Joi.string().email().required().messages({
+    'string.empty': `email cannot be empty`,
+    'any.required': `email is a required field`
+  }),
+  password: Joi.string().alphanum().min(8).required().messages({
+    'string.min': `password should be at least 8 characters`,
+    'any.required': `password is a required field`
+  }),
+});
+
+
+module.exports = validateSchema;
